[compiler-v2] Flush the abstract stack in FF generator if values are still used
If a value is left on the stack from a previous call and then passed on to a subsequent call, the abstract stack in `function_generator.rs` tries to avoid to store it in a local but instead to directly consume from the stack. However, if that value is used afterwards again, it needs to be poped from the stack and stored in a local. In this PR `abstract_push_args` is modified to achieve this.

The 1st commit demonstrates the problem in test case `flush_if_alive.move`. The 2nd commit fixes the issue.