[compiler-v2] Fix a bug in auto reference mode of field selection inn bytecode generator
For a nested field selection as in `(x.f).g`, the inner field selection must be compiled in reference mode, creating a reference for `x.f` instead the value of the field. However, the reference mode was applied to all inner expressions recursively, whether field selection or not. For example, in `borrow_global<T>(x.foo).bar`, `x.foo` was also compiled in reference mode, leading to wrongly typed bytecode. In this PR, we ensure to enter reference mode only if the operand of field selection is again a field selection.

The 1st commit demonstrates the bug, the 2nd fixes it.