[compiler-v2] Do not propagate reference mode to nested expressions in bytecode generator
For a nested field selection as in `(x.f).g`, the inner field selection must be compiled in reference mode, creating a reference for `x.f` and for `x`.  However, the reference mode must not be applied to nested expressions. For example, in `borrow_global<T>(x.foo).bar`, `x.foo` must not be compiled in reference mode. This PR introdues a way to escape reference mode for nested expressions and applies it where appropriate in the bytecode generator.

The 1st commit demonstrates the bug, the 2nd fixes it. The test case file has been renamed to match the actual problem being fixed as it became clearer during debugging.